requirements-test.txt
pytest-mock==3.12.0
pytest-homeassistant-custom-component==0.13.88
